REFACTOR: removed leading underscore so things autoappear in the docs
